restore legacy analysis script
