+++ Python/analysis/script_after_session2.py
@@ -1,3 +1,4 @@
+# ruff: noqa
 
 from eyehead.functions import *
